+++ NEWS.md
@@ -2,8 +2,9 @@
 ### acefile 0.4.2-dev
 
 -   Decode all currently known NT file attributes.
--   Handle archives with multiple different passwords gracefully in CLI.
 -   Avoid rare IndexError when decompressing malformed archives.
+-   Handle archives with multiple different passwords gracefully in CLI.
+-   Print archives in ASCII box with a title to improve clarity in CLI.
 
 
 ### acefile 0.4.1 2017-06-24