+++ NEWS.md
@@ -1,4 +1,7 @@
 
+### acefile 0.4.3-dev
+
+
 ### acefile 0.4.2 2017-06-25
 
 -   Decode all currently known NT file attributes.